Moved ES aggs specs in their own dir
closes #280
