+++ fe/src/lib/elasticsearch/aggs/spec/auto_date_histogram.js
