+++ fe/src/lib/elasticsearch/aggs/spec/avg.js
