+++ fe/src/lib/elasticsearch/aggs/spec/boxplot.js
