+++ fe/src/lib/elasticsearch/aggs/spec/date_histogram.js
