+++ fe/src/lib/elasticsearch/aggs/spec/date_range.js
