+++ fe/src/lib/elasticsearch/aggs/spec/geo_bounds.js
