+++ fe/src/lib/elasticsearch/aggs/spec/geo_centroid.js
