+++ fe/src/lib/elasticsearch/aggs/spec/geohash_grid.js
