+++ fe/src/lib/elasticsearch/aggs/spec/histogram.js
