+++ fe/src/lib/elasticsearch/aggs/spec/min.js
