+++ fe/src/lib/elasticsearch/aggs/spec/range.js
