+++ fe/src/lib/elasticsearch/aggs/spec/rate.js
