+++ fe/src/lib/elasticsearch/aggs/spec/scripted_metric.js
