+++ fe/src/lib/elasticsearch/aggs/spec/variable_width_histogram.js
