+++ fe/src/lib/elasticsearch/aggs/spec/weighted_avg.js
